requirements.txt
# ==================================
# Core Dependencies
# ==================================
pandas>=2.0.0
numpy>=1.24.0
ccxt>=4.0.0
requests>=2.28.0

# ==================================
# Mudrex SDK (installed from local)
# ==================================
# Installed from ./mudrex-sdk

# ==================================
# Environment & Config
# ==================================
python-dotenv>=1.0.0

# ==================================
# Development & Testing
# ==================================
pytest>=7.0.0
pytest-asyncio>=0.21.0

# ==================================
# Railway / Production
# ==================================
# No extra deps needed, just standard python container
